refactor: moved write logic to impl OutputFile
